requirements.txt
streamlit>=1.35.0
duckduckgo-search>=6.2.0
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.2.0
pandas>=2.2.0
python-dotenv>=1.0.1
phonenumbers>=8.13.39
validators>=0.28.0
SQLAlchemy>=2.0.32
psycopg[binary]>=3.2.1
